requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
dnspython==2.2.0
fastapi==0.73.0
Flask==2.0.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.5
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
LoginRadius-v2==11.3.0
MarkupSafe==2.0.1
pbkdf2==1.3
pycparser==2.21
pydantic==1.9.0
pymongo==4.0.1
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
Werkzeug==2.0.2
